Refactored firstyearinfo to be responsive and better written
Table of contents with a bunch of links that scroll to the respective article/guide. On my PC for some reason it would not apply the css properly and make it seem like it is not responsive and not handle the overflows properly. It needs a styling overhaul as all other components