Created today's menu component and put it in restaurant page
The function that determines whether to display this lunch dinner or next day's lunch might need some optimization, i have written a test for personal use where do i upload it? Also it includes an interval every 30 seconds the id of which i log for anyone that wants to clearInterval() it.