More Constants for configurability
In the constants.js some constants were added for easier modifications.

The services page can now be modified easily by changes in the servicesData object(custom ordering with indexes).

Library Hours can now be modified by changing the library_opening_hours constant.

URLs in the AboutPage were extracted to a constant for easier changes in the future.

The restaurant schedule had hard coded values for the menu opening and closing hours, now a constant is used.