Implemented Language Picker and some refactors
For the language picker it was necessary to reload with clear cache for the changes to happen. Also there is a problem with the Announcements Icon we should replace it. For refactors i did some renames,extracted more components and grouped components in folders. More renames are needed as some things are not named properly. This feature needs some QA to make sure weird bugs are not present