Implement holiday break logic and next open date calculation for restaurant schedule
Added logic to check for holiday breaks using ClosedDueToHoliday.jsx and calculate the next open date after the break.
Added hardcoded Easter holiday dates for each year until 2031, summer break dates, and Christmas break dates until 2031 (keep in mind that summer dates are not standard and vary each year).
In the RestaurantPage.jsx, we check if it is a holiday break. If it is, there is no reason to preview today's menu, only the menu for all dates, as the restaurant is closed.